+++ source/state_representation/test/tests/test_parameter.cpp
@@ -246,9 +246,14 @@ TYPED_TEST_P(ParameterTest, MakeShared) {
     EXPECT_EQ(param->get_parameter_type(), std::get<1>(test_case));
     EXPECT_EQ(param->get_parameter_state_type(), std::get<2>(test_case));
     expect_values_equal(param->get_value(), std::get<0>(test_case));
+    copy_parameter_value(param, param_interface);
+    expect_values_equal(param_interface->template get_parameter_value<TypeParam>(), std::get<0>(test_case));
+  
     auto empty_param = make_shared_parameter<TypeParam>("test");
     EXPECT_TRUE(empty_param->is_empty());
     EXPECT_FALSE(*empty_param);
+
+    EXPECT_THROW(copy_parameter_value(empty_param, param_interface), exceptions::EmptyStateException);
   }
 }